requirements.txt
# Core LLM & Agent Framework
langgraph==0.2.45
langchain==0.3.7
langchain-google-genai==2.0.5
google-generativeai==0.8.3

# Vector Database & Embeddings
# chromadb==0.5.23  # Requires C++ build tools - skip for testing
# sentence-transformers==3.3.1
qdrant-client==1.7.3

# API Framework
fastapi==0.115.6
uvicorn[standard]==0.34.0
pydantic==2.10.3
python-multipart==0.0.20

# Firebase
firebase-admin==6.6.0

# Data Processing
pandas>=2.0.0
numpy<2.0.0,>=1.26.0
scipy>=1.10.0

# Utilities
python-dotenv==1.0.1
httpx==0.28.1
tenacity==9.0.0
httplib2>=0.22.0

# Supabase Integration
supabase>=2.0.0

# Testing
pytest==8.3.4
pytest-asyncio==0.25.2
